Separate lottery rpc method to own file
